feat: add timestamps helper for consistent timestamp handling in tasks and users schemas
