+++ src/users/users.schema.ts
@@ -1,6 +1,7 @@
-import { pgTable, text, uuid, timestamp } from 'drizzle-orm/pg-core';
+import { pgTable, text, uuid } from 'drizzle-orm/pg-core';
 import { relations } from 'drizzle-orm';
 import { tasks } from '../tasks/tasks.schema';
+import { timestamps } from '../drizzle/helpers/timestamps.schema';
 
 export const users = pgTable('users', {
   id: uuid('id').primaryKey().defaultRandom(),
@@ -9,16 +10,7 @@ export const users = pgTable('users', {
   email: text('email').notNull().unique(),
   password: text('password').notNull(),
   refreshToken: text('refreshToken'),
-  created_at: timestamp('created_at', {
-    mode: 'date',
-    precision: 0,
-  })
-    .notNull()
-    .defaultNow(),
-  updated_at: timestamp('updated_at', { mode: 'date', precision: 0 })
-    .notNull()
-    .defaultNow()
-    .$onUpdate(() => new Date()),
+  ...timestamps,
 });
 
 export const usersRelations = relations(users, ({ many }) => ({
